Status of actuators now saved
SharedPreferences has been used to save the data of the actuators. The save method is called in the reasoners whenever the users turns on or off an actuator. Couldn't get sharedpreferences working with an ArrayList that contained both types of objects, so that has been walked back to there being a seperate arraylist for light's and lock's again. The loading of the data is called onCreate